requirements.txt
Flask==2.0.1
requests==2.25.1
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
pyairtable==1.0.0
gunicorn==20.1.0  # Add gunicorn for production server